fix: webapp supports setting dependencies
