+++ docs/index.md
@@ -8,6 +8,7 @@ cli
 fixtures
 checks
 families
+webapp
 ```
 
 ```{include} ../README.md
